新增 Force

+++ autoFixEFI.ps1
@@ -1,7 +1,8 @@
 function autoFixEFI {
     param (
         [Parameter(Position = 0, ParameterSetName = "", Mandatory=$true)]
-        [string] $DriveLetter
+        [string] $DriveLetter,
+        [switch] $Force
     )
     $Dri=(Get-Partition -DriveLetter:$DriveLetter)
     
@@ -25,7 +26,7 @@ function autoFixEFI {
         Write-Host " ($($DriveLetter):) " -ForegroundColor:Yellow -NoNewline
         Write-Host "壓縮300M並建立EFI分區" 
         $response = Read-Host "  沒有異議或看不懂，請輸入Y (Y/N) ";
-        if ($response -ne "Y" -or $response -ne "Y") { Write-Host "使用者中斷" -ForegroundColor:Red; return; }
+        if ($response -ne "Y" -or $response -ne "Y" -or $Force) { Write-Host "使用者中斷" -ForegroundColor:Red; return; }
         $Dri|Resize-Partition -Size:($Dri.size-$EfiSize)
         $EFI = (($Dri|New-Partition -Size:($EfiSize) -GptType:$EFI_ID)|Format-Volume)|Get-Partition
     }
@@ -38,7 +39,7 @@ function autoFixEFI {
     Write-Host "寫入EFI分區" -NoNewline
     Write-Host " (磁碟:$($EFI.DiskNumber), 分區:$($EFI.PartitionNumber)) " -ForegroundColor:Yellow
     $response = Read-Host "  沒有異議或看不懂，請輸入Y (Y/N) "
-    if (($response -ne "Y") -or ($response -ne "Y")) { Write-Host "使用者中斷" -ForegroundColor:Red; return; }
+    if (($response -ne "Y") -or ($response -ne "Y" -or $Force)) { Write-Host "使用者中斷" -ForegroundColor:Red; return; }
     
     # 新增EFI磁碟代號
     if(!$EFI.DriveLetter){ $EFI|Set-Partition -NewDriveLetter:$EFI_Letter; $EFI=$EFI|Get-Partition; }
